moved the two remote auth tests over to the system auth tests
